requirements.txt
Django==1.9
django-reversion==2.0.5
Pillow==3.4.2
psycopg2==2.6.2
wheel==0.24.0
